% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_s2s_files.R
\name{generate_output_dir}
\alias{generate_output_dir}
\title{generate output dir}
\usage{
generate_output_dir(output_dir_location)
}
\arguments{
\item{output_dir_location}{Path to the location you want the directories to be generated}
}
\value{
the path to the directory with the current data
}
\description{
This function generates two dirs, one 'analysis_output' and within this a directory
with the current date
}
