% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGplots.R
\name{GOenr_simplified}
\alias{GOenr_simplified}
\title{GOenr_simplified}
\usage{
GOenr_simplified(
  gene_list,
  output_file,
  res,
  dds,
  fill_annotation = "disease",
  x_axis_annotation = "timepoint",
  topn_genes = 9,
  topn_GOterms = 20,
  gene_background_list = "NA",
  qval_cutoff = 0.05,
  gene_ID = "SYMBOL",
  GO_ontology = "BP",
  simplify_value = 0.7
)
}
\arguments{
\item{gene_list}{list of genes (DEGS)}

\item{output_file}{name of the output (pdf) file}

\item{res}{DESeq2 result of differential testing, used for selecting most significant genes from gene_list for plotting}

\item{dds}{deseq2 object, used for grabbing count data of genes of GO-term}

\item{fill_annotation}{metadata column deseq2 object used for count plotting fill}

\item{x_axis_annotation}{metadata column deseq2 object used for count seperation x-axis}

\item{topn_genes}{n genes plotted per GO-term}

\item{topn_GOterms}{n GO-terms plotted}

\item{gene_background_list}{optional gene background}

\item{simplify_value}{value wich is used to simplify go terms, smaller number mean less GO-concationation.}
}
\description{
This function runs GO enrichment using clusterprofilea nd simplify on the outcome
}
