% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGplots.R
\name{GSEA_mysigdb}
\alias{GSEA_mysigdb}
\title{GSEA_mysigdb}
\usage{
GSEA_mysigdb(res, output_dir, ncores = 4)
}
\arguments{
\item{res}{deseq2 result object with gene symbols as identifier}

\item{output_dir}{name of the output directory}

\item{ncores}{ammount of cores to use, on multicore machines you might run into memory issues when you use to many cores.}
}
\description{
This function runs GSEA enrichment using clusterprofiler
}
