% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_s2s_files.R
\name{load_s2s_counts_data}
\alias{load_s2s_counts_data}
\title{load seq2science countable}
\usage{
load_s2s_counts_data(seq2science_dir)
}
\arguments{
\item{seq2science_dir}{Path to the directory where the seq2science results are located}
}
\value{
counts in a dataframe
}
\description{
This function loads the seq2science counts.tsv file
}
