% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_s2s_files.R
\name{load_s2s_sample_file}
\alias{load_s2s_sample_file}
\title{load seq2science sample file}
\usage{
load_s2s_sample_file(seq2science_dir)
}
\arguments{
\item{seq2science_dir}{Path to the directory where the seq2science results are located}
}
\value{
sample file as a dataframe
}
\description{
This function loads the seq2science samples.tsv file
}
