% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preDEGplots.R
\name{plot_dist_dds}
\alias{plot_dist_dds}
\title{plot sample distances}
\usage{
plot_dist_dds(
  normalized_deseq_int,
  output_file,
  annotation1 = NA,
  annotation2 = NA
)
}
\arguments{
\item{normalized_deseq_int}{rlog or vst normalized deseq2 object}

\item{output_file}{name of the output (pdf) file}

\item{annotation1}{potential name of metadata1 in the dds object plotted as annotation}

\item{annotation2}{potential name of metadata2 in the dds object plotted as annotation}
}
\description{
This function plots the sample distance of samples within a deseq2 object
}
