% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preDEGplots.R
\name{plotPCA_dds}
\alias{plotPCA_dds}
\title{plot PCA of samples}
\usage{
plotPCA_dds(
  object,
  intgroup = c("condition1", "condition2"),
  output_dir,
  run_go_enrichment = TRUE,
  Genenames_keyType = "SYMBOL",
  go_term_n_genes = 100,
  n_PCA_top_genes = 20,
  ntop = 1000,
  PC_cutoff = 5,
  gene_background = "not_provided",
  filename = "PCA_"
)
}
\arguments{
\item{object}{deseq2 object}

\item{intgroup}{potential metadata columns, e.g. c('genotype', 'timepoint')}

\item{output_dir}{directory where pdf files will be generated}

\item{run_go_enrichment}{True or False wether to run GO_enrichment}

\item{Genenames_keyType}{genename type, currently only 'SYMBOL'is supported}

\item{go_term_n_genes}{how many genes top genes from the PC should be inc. in the GO enrichment}

\item{n_PCA_top_genes}{how many genes from the PC should be plotted in a table}

\item{ntop}{how many highly variable genes should be inputted in the PCA}

\item{PC_cutoff}{minimal ammount of variance explained by PC for it to be plotted}

\item{gene_background}{optional list of background gene names}

\item{filename}{start of the output filename}
}
\description{
This function plots the sample distance of samples within a deseq2 object
}
