% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGplots.R
\name{GOenr_simplified}
\alias{GOenr_simplified}
\title{GOenr_simplified}
\usage{
GOenr_simplified(
  gene_list,
  output_file,
  gene_background_list = "NA",
  qval_cutoff = 0.05,
  gene_ID = "SYMBOL",
  GO_ontology = "BP",
  simplify_value = 0.7
)
}
\arguments{
\item{gene_list}{list of genes (DEGS)}

\item{output_file}{name of the output (pdf) file}

\item{annotation1}{potential name of metadata1 in the dds object plotted as annotation}

\item{annotation2}{potential name of metadata2 in the dds object plotted as annotation}

\item{costum_sample_order}{list of all sample names in the desired order}
}
\description{
This function runs GO enrichment using clusterprofilea nd simplify on the outcome
}
